+++ src/assets/cet/company.py
@@ -8,6 +8,7 @@ This module contains:
 from __future__ import annotations
 
 import json
+import os
 from pathlib import Path
 
 import pandas as pd
