+++ src/enrichers/naics/fiscal/enricher.py
@@ -7,7 +7,7 @@ from typing import Any
 import pandas as pd
 from loguru import logger
 
-from ...config.loader import get_config
+from ....config.loader import get_config
 from .strategies import (
     AgencyDefaultsStrategy,
     NAICSEnrichmentResult,
