+++ src/loaders/__init__.py
@@ -1,7 +1,6 @@
 """Data loading modules for Neo4j database operations."""
 
-from .neo4j_client import LoadMetrics, Neo4jClient, Neo4jConfig
-from .patent_loader import PatentLoader, PatentLoaderConfig
+from .neo4j import LoadMetrics, Neo4jClient, Neo4jConfig, PatentLoader, PatentLoaderConfig
 
 
 __all__ = ["Neo4jClient", "Neo4jConfig", "LoadMetrics", "PatentLoader", "PatentLoaderConfig"]
