+++ src/transformers/fiscal/roi.py
@@ -17,8 +17,8 @@ from typing import Any
 import pandas as pd
 from loguru import logger
 
-from ..config.loader import get_config
-from ..models.fiscal_models import FiscalReturnSummary
+from ...config.loader import get_config
+from ...models.fiscal_models import FiscalReturnSummary
 
 
 @dataclass
