+++ src/transformers/fiscal/taxes.py
@@ -15,7 +15,7 @@ from typing import Any
 import pandas as pd
 from loguru import logger
 
-from ..config.loader import get_config
+from ...config.loader import get_config
 
 
 @dataclass
